refactor!: rename tck to test-kit since it's more accurate
BREAKING CHANGE: classes in org.smooks.tck relocated to org.smooks.testkit; Maven artifact smooks-tck relocated to smooks-test-kit
