+++ core/pom.xml
@@ -41,7 +41,7 @@
                         </goals>
                         <configuration>
                             <includes>
-                                <include>**/tck/**</include>
+                                <include>**/test-kit/**</include>
                             </includes>
                         </configuration>
                     </execution>
