+++ core/src/test/java/org/smooks/engine/delivery/DomModelCreatorTestCase.java
@@ -48,7 +48,7 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.smooks.tck.Assertions.compareCharStreams;
+import static org.smooks.testkit.Assertions.compareCharStreams;
 
 import org.smooks.Smooks;
 import org.smooks.api.ExecutionContext;
