+++ core/src/test/java/org/smooks/engine/delivery/dom/SmooksVisitorPhaseTestCase.java
@@ -64,7 +64,7 @@ import java.util.List;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertInstanceOf;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.smooks.tck.Assertions.compareCharStreams;
+import static org.smooks.testkit.Assertions.compareCharStreams;
 
 /**
  * Tests to make sure the phase annotations work properly.
