+++ core/src/test/java/org/smooks/testkit/delivery/dom/MockContentDeliveryConfig.java
@@ -1,69 +1,88 @@
 /*-
  * ========================LICENSE_START=================================
- * Test Kit
+ * Core
  * %%
  * Copyright (C) 2020 Smooks
  * %%
  * Licensed under the terms of the Apache License Version 2.0, or
  * the GNU Lesser General Public License version 3.0 or later.
- *
+ * 
  * SPDX-License-Identifier: Apache-2.0 OR LGPL-3.0-or-later
- *
+ * 
  * ======================================================================
- *
+ * 
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * 
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *
+ * 
  * ======================================================================
- *
+ * 
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
  * License as published by the Free Software Foundation; either
  * version 3 of the License, or (at your option) any later version.
- *
+ * 
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
  * Lesser General Public License for more details.
- *
+ * 
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  * =========================LICENSE_END==================================
  */
-package org.smooks.tck;
+package org.smooks.testkit.delivery.dom;
 
-import org.hsqldb.server.ServerConstants;
-import org.junit.jupiter.api.Test;
+import org.smooks.engine.delivery.ContentHandlerBindingIndex;
+import org.smooks.engine.delivery.dom.DOMContentDeliveryConfig;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import java.util.ArrayList;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
 
 /**
- * Unit test for {@link HsqlServer}.
+ * Mock ContentDeliveryConfig for DOM.
  *
- * @author <a href="mailto:danielbevenius@gmail.com">Daniel Bevenius</a>
+ * @author tfennelly
  */
-public class HsqlServerTestCase {
-    @Test
-    public void startStop() throws Exception {
-        for (int i = 0; i < 50; i++) {
-            startStopCycle();
-        }
-    }
+@SuppressWarnings({ "unchecked", "unused" })
+public class MockContentDeliveryConfig extends DOMContentDeliveryConfig {
+  private final Map objectsHash = new LinkedHashMap();
 
-    private void startStopCycle() throws Exception {
-        HsqlServer hsqlServer = new HsqlServer(1999);
-        hsqlServer.stop();
-        assertEquals(ServerConstants.SERVER_STATE_SHUTDOWN, hsqlServer.getState());
-    }
+  public MockContentDeliveryConfig() {
+    setResourceConfigs(new LinkedHashMap<>());
+    setAssemblyVisitBeforeIndex(new ContentHandlerBindingIndex());
+    setAssemblyVisitAfterIndex(new ContentHandlerBindingIndex());
+    setProcessingVisitBeforeIndex(new ContentHandlerBindingIndex());
+    setProcessingVisitAfterIndex(new ContentHandlerBindingIndex());
+    setSerializerVisitorIndex(new ContentHandlerBindingIndex());
+    setPostFragmentLifecycleIndex(new ContentHandlerBindingIndex());
+  }
 
+  /* (non-Javadoc)
+   * @see org.smooks.engine.delivery.ContentDeliveryConfig#getObjects(java.lang.String)
+   */
+  public List getObjects(String selector) {
+    return (List) objectsHash.get(selector);
+  }
+
+  public void addObject(String selector, Object object) {
+    List objects = (List) objectsHash.get(selector);
+
+    if (objects == null) {
+      objects = new ArrayList<>();
+      objectsHash.put(selector, objects);
+    }
+    objects.add(object);
+  }
 }
