+++ pom.xml
@@ -77,7 +77,7 @@
         <module>commons</module>
         <module>core</module>
         <module>management</module>
-        <module>tck</module>
+        <module>test-kit</module>
         <module>benchmark</module>
     </modules>
