+++ test-kit/src/main/java/org/smooks/testkit/StdoutToLog4jFilter.java
@@ -40,7 +40,7 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  * =========================LICENSE_END==================================
  */
-package org.smooks.tck;
+package org.smooks.testkit;
 
 import org.slf4j.Logger;
